pyproject.toml
[project]
name = "btc-options-onchain-engine"
version = "0.1.0"
description = "Production-grade BTC options + onchain inference engine for regime detection and distributional forecasting"
authors = [
    {name = "Quant Team", email = "quant@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core data science
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    
    # Advanced ML
    "lightgbm>=4.1.0",
    "statsmodels>=0.14.0",
    
    # Database
    "duckdb>=0.9.0",
    "sqlalchemy>=2.0.0",
    
    # API clients
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    "requests>=2.31.0",
    
    # Data validation
    "pydantic>=2.4.0",
    "pydantic-settings>=2.0.0",
    
    # CLI and orchestration
    "typer>=0.9.0",
    "prefect>=2.14.0",
    "rich>=13.6.0",
    
    # Visualization
    "plotly>=5.17.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    
    # Dashboard
    "streamlit>=1.28.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.1",
    "jinja2>=3.1.2",
    "pyarrow>=14.0.0",
    "fastparquet>=2023.10.0",
    "tenacity>=8.2.3",
    "cachetools>=5.3.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.10.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
    "ipython>=8.16.0",
    "jupyter>=1.0.0",
]

[project.scripts]
btc-engine = "btc_engine.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.10.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=btc_engine --cov-report=term-missing"
